+++ react/components/AppRouter.tsx
@@ -15,15 +15,11 @@ import AddressCreate from './pages/AddressCreate'
 import AddressEdit from './pages/AddressEdit'
 import Menu from './Menu'
 
-interface AppRouterProps {
+interface Props {
   blockDocument?: boolean
 }
 
-class AppRouter extends Component<AppRouterProps> {
-  constructor(props: AppRouterProps) {
-    super(props)
-  }
-
+class AppRouter extends Component<Props> {
   public state = { defaultPath: '' }
 
   private handleDefaultPath = (defaultPath: string) => {
@@ -50,10 +46,8 @@ class AppRouter extends Component<AppRouterProps> {
     component,
   }: {
     path: string
-   component: ComponentClass<void, unknown> | FC
-
-  }) => <Route exact key={path} path={path} component={component}/>
-
+    component: ComponentClass<void, unknown> | FC
+  }) => <Route exact key={path} path={path} component={component} />
 
   public render() {
     const routes = [
@@ -61,7 +55,12 @@ class AppRouter extends Component<AppRouterProps> {
       { path: '/addresses/new', component: AddressCreate },
       { path: '/addresses/edit/:id', component: AddressEdit },
       { path: '/profile', component: Profile },
-      { path: '/profile/edit', component: ()=><ProfileEdit blockDocument={this.props.blockDocument}/> },
+      {
+        path: '/profile/edit',
+        component: () => (
+          <ProfileEdit blockDocument={this.props.blockDocument} />
+        ),
+      },
     ]
 
     return (